requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pip==21.3.1
psycopg2==2.9.5
pytz==2022.7.1
requests==2.28.2
setuptools==60.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wheel==0.37.1

django-mptt~=0.14.0